requirements.txt
python-dotenv
requests
pandas
psycopg2-binary
boto3
minio
pymssql